Added user service layer to get user plan details and also validate user details

+++ subscription-service/src/main/java/ai/sahaj/subscription/controller/UserPlanController.java
@@ -18,7 +18,7 @@ public class UserPlanController {
 
 
   @GetMapping("/{userId}")
-  public UserPlanResponse getUserPlanDetailsForUserId(@PathVariable("userId") Long userId) {
+  public UserPlanResponse getUserPlanDetailsForUserId(@PathVariable("userId") Integer userId) {
     return UserPlanResponse.from(userPlanService.getUserPlanDetails(userId));
   }
