+++ subscription-service/src/main/java/ai/sahaj/subscription/dto/UserPlanResponse.java
@@ -5,7 +5,7 @@ import ai.sahaj.subscription.domain.UserPlanDO;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record UserPlanResponse(Long userId, Integer planId, String planName, BigDecimal dataBalance,
+public record UserPlanResponse(Integer userId, Integer planId, String planName, BigDecimal dataBalance,
                                LocalDate planStartedOn) {
   public static UserPlanResponse from(UserPlanDO userPlanDO) {
     return new UserPlanResponse(userPlanDO.userId(), userPlanDO.planId(), userPlanDO.planName(), userPlanDO.getDataBalance(),
